requirements.txt
langchain
langchain-community
langchain-groq
python-dotenv
crewai
streamlit
crewai-tools 

-e .